backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
bcrypt==4.0.1
passlib==1.7.4
python-multipart
python-dotenv
requests
groq
pandas
openpyxl